+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 2.3.0 - Unreleased
 ### Added
 - Added the ability to dynamically get templates with the `craft.blitz.getTemplate` template tag and the `TemplatesController::actionGet` controller action ([#129](https://github.com/putyourlightson/craft-blitz/issues/129)).
-- Added unique zone IDs for each site in the Cloudflare reverse proxy purger settings.
+- Added unique zone IDs for each site to the Cloudflare reverse proxy purger settings.
+- Added warm cache delay setting to the Cloudflare reverse proxy purger settings.
 - Added unit tests to the plugin.
 
 ### Fixed